fix(#9): temporary fix
After multiple researches and trials, there seems to be a bug when we enter names that are either too long or containing at date after 2022... We should create an issue in the appropriate repository but I still don't understand where the error really comes from. I have heard a bug about that 2022-... parsed into int goes above the max value but I don't think it is related.

A quick fix is to replace "-" with "_".

Closes gh-9